Backend: Kafka + Mail Notification. Remain commit. Done.

+++ backend/orders/src/main/java/com/olomsky/orders/dto/OrderResponse.java
@@ -1,8 +1,10 @@
 package com.olomsky.orders.dto;
 
+import java.math.BigDecimal;
+
 public record OrderResponse(
     Long id,
     String orderNumber,
     String skuCode,
-    Double price,
+    BigDecimal price,
     Integer quantity) {}
